Basic random bot
This version includes a basic random bot which randomly chooses a ball to move and moves it to any random spot.

Have also started to create a function which identifeis "partial long rows" (ie 2 or more in a row out of 5 in a row)

However it doesn't work because if some of the 5 in a row are occupied with balls of a different colour, it still recognises this as a partial long row.